Move python tests next to the files they test
