+++ source/python/.coveragerc
@@ -2,7 +2,9 @@
 include =
     */neuropod/*
     */_neuropod_native_bootstrap/*
-omit = *tests*
+omit =
+    */test/*
+    */tests/*
 concurrency = multiprocessing
 data_file = /tmp/neuropod_py_coverage/.coverage
